Move engine types to engine mod
