+++ scrypto/src/component/system.rs
@@ -2,9 +2,7 @@ use crate::abi::BlueprintAbi;
 use crate::buffer::*;
 use crate::component::*;
 use crate::core::Runtime;
-use crate::core::ScryptoRENode;
-use crate::engine::types::RENodeId;
-use crate::engine::{api::*, call_engine};
+use crate::engine::{api::*, types::*, utils::*};
 use sbor::rust::collections::*;
 use sbor::rust::string::String;
 use sbor::rust::string::ToString;
