+++ scrypto/src/core/logger.rs
@@ -1,7 +1,6 @@
 use sbor::rust::string::String;
 
-use crate::core::*;
-use crate::engine::{api::*, call_engine};
+use crate::engine::{api::*, types::*, utils::*};
 
 /// A utility for logging messages.
 #[derive(Debug)]
