+++ scrypto/src/core/mod.rs
@@ -2,8 +2,6 @@ mod actor;
 mod blob;
 mod data;
 mod expression;
-mod invocation;
-mod level;
 mod logger;
 mod network;
 mod runtime;
@@ -12,8 +10,6 @@ pub use actor::ScryptoActor;
 pub use blob::*;
 pub use data::*;
 pub use expression::*;
-pub use invocation::*;
-pub use level::Level;
 pub use logger::Logger;
 pub use network::{NetworkDefinition, ParseNetworkError};
 pub use runtime::{
