+++ scrypto/src/core/transaction.rs
@@ -1,5 +1,5 @@
 use crate::crypto::*;
-use crate::engine::{api::*, call_engine};
+use crate::engine::{api::*, types::*, utils::*};
 
 /// The transaction context at runtime.
 #[derive(Debug)]
