+++ scrypto/src/engine/types/ids.rs
@@ -1,0 +1,11 @@
+use super::*;
+
+pub type LockHandle = u32;
+pub type AuthZoneId = u32;
+pub type BucketId = u32;
+pub type ProofId = u32;
+
+pub type ComponentId = (Hash, u32);
+pub type KeyValueStoreId = (Hash, u32);
+pub type NonFungibleStoreId = (Hash, u32);
+pub type VaultId = (Hash, u32);
