+++ scrypto/src/engine/types/third_party.rs
@@ -1,0 +1,12 @@
+pub use sbor::rust::fmt;
+pub use sbor::rust::string::*;
+pub use sbor::rust::vec::Vec;
+pub use sbor::*;
+pub use sbor::*;
+pub use sbor::*;
+pub use strum::*;
+
+pub use crate::component::{ComponentAddress, PackageAddress};
+pub use crate::crypto::Hash;
+pub use crate::resource::NonFungibleId;
+pub use crate::resource::ResourceAddress;
