+++ scrypto/src/macros.rs
@@ -144,7 +144,7 @@ macro_rules! args_from_bytes_vec {
 #[macro_export]
 macro_rules! error {
     ($($args: expr),+) => {{
-        ::scrypto::core::Logger::log(scrypto::core::Level::Error, ::sbor::rust::format!($($args),+));
+        ::scrypto::core::Logger::log(scrypto::engine::types::Level::Error, ::sbor::rust::format!($($args),+));
     }};
 }
 
@@ -159,7 +159,7 @@ macro_rules! error {
 #[macro_export]
 macro_rules! warn {
     ($($args: expr),+) => {{
-        ::scrypto::core::Logger::log(scrypto::core::Level::Warn, ::sbor::rust::format!($($args),+));
+        ::scrypto::core::Logger::log(scrypto::engine::types::Level::Warn, ::sbor::rust::format!($($args),+));
     }};
 }
 
@@ -174,7 +174,7 @@ macro_rules! warn {
 #[macro_export]
 macro_rules! info {
     ($($args: expr),+) => {{
-        ::scrypto::core::Logger::log(scrypto::core::Level::Info, ::sbor::rust::format!($($args),+));
+        ::scrypto::core::Logger::log(scrypto::engine::types::Level::Info, ::sbor::rust::format!($($args),+));
     }};
 }
 
@@ -189,7 +189,7 @@ macro_rules! info {
 #[macro_export]
 macro_rules! debug {
     ($($args: expr),+) => {{
-        ::scrypto::core::Logger::log(scrypto::core::Level::Debug, ::sbor::rust::format!($($args),+));
+        ::scrypto::core::Logger::log(scrypto::engine::types::Level::Debug, ::sbor::rust::format!($($args),+));
     }};
 }
 
@@ -204,7 +204,7 @@ macro_rules! debug {
 #[macro_export]
 macro_rules! trace {
     ($($args: expr),+) => {{
-        ::scrypto::core::Logger::log(scrypto::core::Level::Trace, ::sbor::rust::format!($($args),+));
+        ::scrypto::core::Logger::log(scrypto::engine::types::Level::Trace, ::sbor::rust::format!($($args),+));
     }};
 }
