+++ scrypto/src/resource/auth_zone.rs
@@ -3,9 +3,7 @@ use sbor::rust::string::ToString;
 use sbor::rust::vec::Vec;
 use sbor::*;
 
-use crate::core::{AuthZoneMethod, NativeMethodIdent, Receiver};
-use crate::engine::types::RENodeId;
-use crate::engine::{api::*, call_engine};
+use crate::engine::{api::*, types::*, utils::*};
 use crate::math::Decimal;
 use crate::native_functions;
 use crate::resource::*;
