+++ scrypto/src/resource/proof.rs
@@ -6,9 +6,7 @@ use sbor::rust::vec::Vec;
 use sbor::*;
 
 use crate::abi::*;
-use crate::core::{ProofMethod, Receiver};
-use crate::engine::types::RENodeId;
-use crate::engine::{api::*, call_engine, types::ProofId};
+use crate::engine::{api::*, types::*, utils::*};
 use crate::math::*;
 use crate::misc::*;
 use crate::native_functions;
