+++ scrypto/src/resource/resource_manager.rs
@@ -8,10 +8,7 @@ use sbor::*;
 use crate::abi::*;
 use crate::address::*;
 use crate::buffer::scrypto_encode;
-use crate::core::NativeMethodIdent;
-use crate::core::{Receiver, ResourceManagerMethod};
-use crate::engine::types::{GlobalAddress, RENodeId};
-use crate::engine::{api::*, call_engine};
+use crate::engine::{api::*, types::*, utils::*};
 use crate::math::*;
 use crate::misc::*;
 use crate::native_functions;
